requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0

# Validation
email-validator==2.1.0
python-dateutil==2.8.2

# Notifications
twilio==8.11.0
sendgrid==6.11.0

# WebSocket
python-socketio==5.11.0
eventlet==0.35.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Utilities
python-decouple==3.8
gunicorn==21.2.0